% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteracDiagnosis-package.R
\docType{package}
\name{InteracDiagnosis}
\alias{InteracDiagnosis}
\alias{InteracDiagnosis-package}
\title{InteracDiagnosis}
\description{A Shiny tool for visualising how adding or omitting covariates affects parameter estimates, focusing on changes in the coefficient of interest, along with theoretical and bootstrap standard errors, partial (added-variable) plots, and model-comparison statistics (AIC, BIC, and likelihood-ratio tests).
}
\author{
\strong{Maintainer}: Jianyue \email{jianyue.bai@mail.utoronto.ca}

Other contributors:
\itemize{
  \item Selenge [contributor]
}

}
